Removed some files from tracking, added .gitignore and Makefile.
We do not need logs nor the game save data to be tracked.

Also, I needed to add a Makefile because I apparently need to redo the
World.exe file every time I make a change to the scripts. This is a bit
tedius, so I am going to have Make keep track of what I need to do.