Adding more spawn to Accursed Maze dungeon.
There was an empty subroom where all others had had at least one spawn
in them. This adds to that room to ensure that all are filled.

There is a parameter that I am not sure of in the file that I believe is
the spawner's spawn id. I do not know whether they have to be unique and
whether anything will go bad if they are not.