Fixing wording on minimum stats for equipping some items.
The previous wording made absolutely no sense, so I fixed it so that
it's grammatically correct.

There are some items that have other requirements than just skill, such
as the knight and death knight spellbooks that have karma requirements
that are not being given to the end user if they attempt to equip it
with the required skill level. I will make a new story for that.