Removed skill check that sets all skills to zero.
Removed the bane of me playing so far. This skill check thing that looks
at one's skill cap and decides to zero out skills is awful. I am very
glad that I was able to figure it out so quickly.

There is a lot of bad looking code in PlayerMobile.cs. Just some code
smells that I would like to get rid of eventually. Nothing is bothering
me enough that I feel like it should be changed immediately, but things
like not using curlies to distinguish when conditionals or loops should
be applied for instance and bad indents are present.

Whoever wrote this stuff I think had a good vision, but they were not
the best at programming.