Updated buy chances to 100%
I am tired of having to double-check multiple NPCs to see if they will
buy the stuff that I am getting when I am killing stuff in dungeons. I
am probably losing out on 100's of GP by throwing items on the ground
because I A) do not have enough money for a house right now, B) can't
use the stuff because it's for a different weapon type (will be fixed in
another ticket), and C) am running out of bank storage for the stuff
that I can reasonably want to save!