Updated Hides to be cuttable from ground.
I had to delete a lot because the IScissorable was on each derivative
class instead of the base abstract class. I am hoping that I do not get
any gripes from this, but I will have to double-check once I try it on
my desktop where I can use mono.

What I did was follow the style of Ore.cs in that we have one method in
the base class that handles what should happen in scissors are used on
any hide type and then only have the code that is different in those
derived types by only having them pass the type of leather that should
be given when the like type of hide is cut.