Updated unlock range to 2.
This is another thing that ticked me off. There are some containers that
are not that easy to figure out which tile exactly it is (Think human
corpses being used as dungeon chests).

This is going to make it so that the player can be up to the number of
tiles away to pull things out of it or open the chest with disable trap
anyway. I really do not feel like I'm breaking much here.