Double-click to unlock containers!
Woohoo! This one unfortunately took some doing, but we're good to go
now.

The problem was how tightly coupled the idea of targeting was to the
logic that was handling unlocking a lock. Once that was removed into the
Lockpick class (as it probably should have been before), then it was
pretty simple.