Added pre-commit config, renamed csproj file.
Going to need to move the csproj file in order for the dotnet-format to
work because it does not like the sln and the csproj being in the same
folder. It gets confused, bless its heart.

I am going to continue trying to get omnisharp to also do its thing on
the files, but I'm having less luck with it.

I DO need to also change the pre-commit hooks to only work on cs, xml,
json, etc. for now because it's trying to mess with .mul files and stuff
which I am not entirely certain is OK.