Moved project files to adhere to new Dotnet project format.
Per Microsoft.com, this is in fitting with their new project style
suggestion. In this way, we have a single sln file at the root of the
parent directory and then all of our projects in sub folders of the src
folder. This keeps things nice and neat.

More importantly for me and why I am doing it now, this should allow for
dotnet-format to be able to work! yay.