Almost done with getting the project setup with formatting.
We're getting really close here. Omnisharp is not working for me as a
formatter, so I am going to start using uncrustify.

I believe I can have uncrustify setup as a project dependency via
pre-commit and have it run through that automagically. This will then
use my extremely, extremely opinionated configuration settings to format
the code. Super excited with what I have seen so far with the
configurations that I have been able to set for it thus far.