First half reformatted.
Ran into an issue trying to reformat HairRestylingDeed and Build.cs.
Need to figure out what's going on there. I am worried because I'm doing
this on a pretty resource constrained laptop, so I do not know if it's
just because this poor thing couldn't handle it, or if there is
something unparsable in the files. Will double-check and retry.