Another two formatted.
Having issues with the same two files again, but I think it's because
this laptop has literally two low-powered cores and the Build.cs file is
16k lines long. I am sure the 4 gigs of ram I have on this thing is
definitely not helping either.

I am going to go to sleep and in the morning turn on my PC, load up this
branch, and let 32 cores, 64 gigs of ram have fun with it. Probably take
a fraction of the time.