Updated files based on feedback from dotnet build.
Trying to do a build has shown that these files had some silly issues.

One of the problems was that the lucky earnings had a bunch of weird
characters between each letter.

The other two just had issues going from implied braces for a
conditional statement with a single line to having braces added. For
some reason, the semicolon got pushed to the next line.