Updating formatting across more files, removed HairRestylingDeed.
For some reason, Uncrustify is having the hardest time on the int[][]
within HairRestylingDeed. So I'm going to remove it for now because it's
really not worth the effort I am going to have to put in to fix it.

So far, I believe it is because of the int[][] as that appears to be
where it is getting stuck as per the stdout that I saw when running
basic uncrustify against the file with my config. I ensured that it was
not because of ints that look like "000" and that it also was not
because of comments at the end of the lines within the declaration of
the array literals.