Trying to get dotnet build to complete successfully.
Running into issues still that uncrustify missed.

Seeing a lot of semicolons not in the right place. Also had an issue
with Serialization where it was putting braces in weird places. I am
hoping that I didn't screw any of that up.

Now I have to remove the pre-aos attributes to a lot of the magical
versions of the base weapon cases. There are a few hundred. Rip.