Added test project to solution, turned off assembly info in proj.
Was having an issue where it was attempting to duplicate assembly info
in the project when I went to try to build it using the normal make that
had been working for me in the past. I had to clear out the obj folder
in order for it to work afterwards.

I probably should include that as an automatic step in my Makefile if
it's not already an option in my compiler to just disregard old builds.