Working on removing scriptcompiler.
This is going to take a lot of work. I'm seeing something like 86
issues.

In the middle of clearing them, I noticed that it appears Dumbass was
using it as a way to get the types of Items. I was able to remove 1.3k
lines from SearchBook that were just insane. Had to make a python script
and just play around with the logic off to the side to even figure out
what was going on.

I was reading some Reddit posts to get more information about the person
that made this stuff, and it was funny because he got really defensive
about people calling him out for things not working, and he said, "I do
this for free, I don't make a cent." I should hope not, dude. No one
should ever get paid for this drivel.