Update items to show a price if they're able to be sold.
This gets a generic item selling object and runs the item against the
algorithm that is used for figuring out the base prices for things.

This might get annoying, so I gotta check whether I want to narrow this
down further than where it is now. This was just the easist thing to
change.

I might instead want to add this to baseweapon, basearmor, basereagent,
basepotion, etc.

On second thought, maybe it will live in Item.