Adding price info to weapon instead of base item.
Item adds it to damn near everything. I saw it on containers, gold, etc.
Not good for sure.

Having this on BaseWeapon makes more sense then if things are
technically able to be sold other than just weapons, armor, etc.

This however is broken because at some point the information is getting
serialized and saved. This will not work with my current game save
unfortunately.